* Replace function parse_float() with a templated version parse_fp; * Make function parse_xyz() templated; Update CMakeLists.txt.

+++ include/engine/parse/xml/camera/projection.hpp
@@ -9,6 +9,6 @@
 namespace engine::parse::xml {
 
 auto parse_projection(rapidxml::xml_node<> const* node) noexcept
-    -> cpp::result<glm::vec3, parse_err>;
+    -> cpp::result<glm::dvec3, parse_err>;
 
 } // namespace engine::parse::xml
